MathExpression reworked, parser.h function prototypes and some implementations
Changed MathExpression from a struct containing a pointer to a instruction tree and the float result to a pointer to a instruciton tree, result shall be obtained with a function.
Added prototypes for instrNode functions: create empty (number and operation versions), create a tree from a string, delete the tree.
Added prototypes for MathExpression functions: create empty, delete the tree (wrapper), parse string to tree (wrapper) and get the result.
Added Implementations for empty nodes and math expression and for math expression wrapper functions